+++ changelog.md
@@ -5,7 +5,6 @@
 - add context menu to invoice table
 - all clients should be loaded, otherwise client selection list all options.
 - Refactor/organize based on [slide for Week 5 Office Hour 1](https://docs.google.com/presentation/d/1pLZ8Zvpyr_myotpvyVYzY5awpUFUO78DE5CnBnVNm7c/edit#slide=id.g117828bad90_0_9)
-- Reemplazar hamburget con speed dial https://mui.com/components/speed-dial/
 - throwing exception here: invoiceslice: thunkAPI.dispatch(beforeUpdate(prevInvoice)); (missing client)
 - if you are in page 10, and filter by client, filters are applied but page is still 10 (unexistant)
 
@@ -18,6 +17,7 @@
   comes first.
 # 2022 03 20
 - cancel buttons should go back to previous page
+- Reeplaced hamburger custom menu with MUI PopupState Menu https://mui.com/components/menus/#popupstate-helper
 # 2022 03 19
 - Add support in invoice store slice to filter and sort invoices.
 - Show more columns in the invoices pages